requirements/dev.txt
tensorflow-gpu
tensorflow
pillow
Augmentor
tqdm
imageio
opencv-python
google_images_download
imgaug